+++ src/gei/Factories/BlobServiceClientFactory.cs
@@ -1,0 +1,11 @@
+using Azure.Storage.Blobs;
+
+namespace OctoshiftCLI.GithubEnterpriseImporter.Factories;
+
+public class BlobServiceClientFactory : IBlobServiceClientFactory
+{
+    public BlobServiceClient Create(string connectionString)
+    {
+        return new BlobServiceClient(connectionString);
+    }
+}
