+++ src/gei/Factories/IAzureApiFactory.cs
@@ -1,0 +1,9 @@
+using OctoshiftCLI.Services;
+
+namespace OctoshiftCLI.GithubEnterpriseImporter.Factories;
+
+public interface IAzureApiFactory
+{
+    AzureApi Create(string azureStorageConnectionString);
+    AzureApi CreateClientNoSsl(string azureStorageConnectionString);
+}
