+++ src/gei/Factories/IBlobServiceClientFactory.cs
@@ -1,0 +1,8 @@
+using Azure.Storage.Blobs;
+
+namespace OctoshiftCLI.GithubEnterpriseImporter.Factories;
+
+public interface IBlobServiceClientFactory
+{
+    public BlobServiceClient Create(string connectionString);
+}
