+++ src/gei/Factories/SecretScanningAlertServiceFactory.cs
@@ -1,7 +1,7 @@
 using OctoshiftCLI.Contracts;
 using OctoshiftCLI.Services;
 
-namespace OctoshiftCLI.GithubEnterpriseImporter;
+namespace OctoshiftCLI.GithubEnterpriseImporter.Factories;
 
 public class SecretScanningAlertServiceFactory
 {
