--- src/gei/IAzureApiFactory.cs
@@ -1,10 +1,0 @@
-using OctoshiftCLI.Services;
-
-namespace OctoshiftCLI.GithubEnterpriseImporter
-{
-    public interface IAzureApiFactory
-    {
-        AzureApi Create(string azureStorageConnectionString);
-        AzureApi CreateClientNoSsl(string azureStorageConnectionString);
-    }
-}
