--- src/gei/IBlobServiceClientFactory.cs
@@ -1,9 +1,0 @@
-using Azure.Storage.Blobs;
-
-namespace OctoshiftCLI.GithubEnterpriseImporter
-{
-    public interface IBlobServiceClientFactory
-    {
-        public BlobServiceClient Create(string connectionString);
-    }
-}
